fix collision

+++ environments/Box.java
@@ -21,7 +21,7 @@ public class Box extends GameComponent implements DestructibleComponent {
   }
 
   public void draw(Graphics2D graphics2d) {
-    graphics2d.drawImage(sprite.getBufferedImage(), x, y, this);
+    graphics2d.drawImage(sprite.getBufferedImage(), (int) x, (int) y, this);
   }
 
   public void update(double deltaTime) {
